+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.12] - 2025-04-23
+
+### Added
+
+- Added `navbar.jsx` with a logo, navigation menu, and sign-up button.
+- Modularized `GetData()` function
+
 ## [0.1.11] - 2025-04-18
 
 ### Added